2player mario, fixed button order.
Ps. This delay thing is going to make me destroy this device. I have no idea if it is C or the board. Will have to watch delays and stuff after every implementation, each delay alters every other delay in the program and can potentially cause an infinite loop of timer interrupts due to stuff running slowly.
It's as if the clock speed changes or some stupid shit like that.